requirements.txt
appnope==0.1.3
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.8
djangorestframework==3.11.2
executing==0.8.3
gunicorn==20.0.4
httpie==3.1.0
idna==3.3
ipython==8.3.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
Markdown==3.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
mysqlclient==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.7.4
PySocks==1.7.1
pytz==2022.1
PyYAML==5.4
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==5.0.1
